+++ src/main/java/META-INF/MANIFEST.MF
@@ -1,3 +1,3 @@
 Manifest-Version: 1.0
-Main-Class: com.example.internProject.InternProjectApplication
+Main-Class: tr.com.orioninc.laborant.InternProjectApplication
