+++ src/main/java/tr/com/orioninc/laborant/Controller/AdminController.java
@@ -1,8 +1,8 @@
-package com.example.internProject.Controller;
+package tr.com.orioninc.laborant.Controller;
 
-import com.example.internProject.Model.Lab;
-import com.example.internProject.Service.AdminService;
-import com.example.internProject.Service.LabService;
+import tr.com.orioninc.laborant.Model.Lab;
+import tr.com.orioninc.laborant.Service.AdminService;
+import tr.com.orioninc.laborant.Service.LabService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
