+++ src/main/java/tr/com/orioninc/laborant/Model/Lab.java
@@ -1,4 +1,4 @@
-package com.example.internProject.Model;
+package tr.com.orioninc.laborant.Model;
 
 import javax.persistence.*;
