+++ src/test/java/tr/com/orioninc/laborant/InternProjectApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.internProject;
+package tr.com.orioninc.laborant;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
